WIP: protocol refactoring
