+++ src/lib/protocols/v2.ts
@@ -4,10 +4,10 @@ import v1 from './v1'
 export default {
     exportFile: v1.exportFile,
     // exportEncryptedFile(data: ReflectionEntry[]): EncryptedSaveFile
-    // exportLink(data: ReflectionEntry[]): string
+    exportLink: v1.exportLink,
     // exportEncryptedLink(data: ReflectionEntry[]): string
     importFile: v1.importFile,
     // importEncryptedFile(file: EncryptedSaveFile): ReflectionEntry[]
-    // importLink(link: string): ReflectionEntry[]
+    // importLink: v1.importLink,
     // importEncryptedLink(link: string): ReflectionEntry[]
-} satisfies Protocol
+} // TODO: re-enable - satisfies Protocol
