+++ src/lib/utils.ts
@@ -1,9 +1,6 @@
-import { base64url } from 'rfc4648'
 import type {
     LifewheelState,
     LifewheelStep,
-    ParsedLink,
-    ProtocolVersion,
     ReflectionEntry,
     ReflectionStep,
     TextStep,
@@ -78,25 +75,22 @@ export function encodeInt16(n: number) {
  * However, I used strings to make the code more readable.
  */
 export function encodeEntryData(data: ReflectionEntry['data']) {
-    const bin = data.map(number => number.toString(2).padStart(4, '0'))
-                
-    return [
-        bin[0] + bin[1],
-        bin[2] + bin[3],
-        bin[4] + bin[5],
-        bin[6] + bin[7],
-    ].map(n => parseInt(n, 2))
+    const bin = data.map((number) => number.toString(2).padStart(4, '0'))
+
+    return [bin[0] + bin[1], bin[2] + bin[3], bin[4] + bin[5], bin[6] + bin[7]].map((n) =>
+        parseInt(n, 2),
+    )
 }
 
 /**
  * Decode entry data back into the original numbers
  */
 export function decodeEntryData(data: number[]) {
-    const bin = data.flatMap(number => {
+    const bin = data.flatMap((number) => {
         const n = number.toString(2).padStart(8, '0')
         return [n.slice(0, 4), n.slice(4)]
     })
-    return bin.map(n => parseInt(n, 2))
+    return bin.map((n) => parseInt(n, 2))
 }
 
 export function decodeInt32(data: Uint8Array) {
@@ -105,32 +99,6 @@ export function decodeInt32(data: Uint8Array) {
 
 export function decodeInt16(data: Uint8Array) {
     return (data[0] << 8) | data[1]
-}
-
-export const formatHeader = ({
-    encrypted,
-    protocolVersion,
-}: {
-    encrypted: boolean
-    protocolVersion: ProtocolVersion
-}) => `${encrypted ? '1' : '0'}e${protocolVersion}p`
-
-export const parseLink = (link: string): ParsedLink => {
-    /**
-     * Link header example: "1e1p" means encryption enabled, and protocol version 1
-     */
-    const match = link.match(/^([10])e(\d+)p/)
-    if (!match) throw new Error(`Invalid header: ${link}`)
-
-    // Remove the header to get the data.
-    const rawData = link.replace(match[0], '')
-    if (!rawData) throw new Error(`Empty data: ${link}`)
-
-    return {
-        encrypted: match[1] === '1',
-        protocolVersion: parseInt(match[2], 10) as ProtocolVersion,
-        data: base64url.parse(rawData),
-    }
 }
 
 /**
